chore(refactor): add test for model
