+++ test/server/models/user.model.test.js
@@ -2,6 +2,7 @@ import bcrypt from 'bcryptjs';
 import { expect, sinon } from '../config';
 import { hashPassword } from '../../../server/services/user.service';
 
+// TODO:
 describe('User Model', () => {
   describe('Hash Password', () => {
     it('should return a hashed password asynchronously', async () => {
